breaking out py utils & updating refs
